requirements.txt
aiojobs==1.0.0
aioredis==1.3.1
alembic==1.7.7
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.3
Deprecated==1.2.13
Faker==13.11.1
fastapi==0.78.0
fastapi-plugins==0.9.0
googleapis-common-protos==1.56.1
graypy==2.1.0
greenlet==1.1.2
grpcio==1.46.1
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-asgi==0.30b1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-instrumentation-httpx==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-sqlalchemy==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-faker==2.0.0
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2022.1
redis==4.3.1
sentry-sdk==1.5.12
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy-datatables==2.0.1
SQLAlchemy-serializer==1.4.1
starlette==0.19.1
tenacity==8.0.1
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
